# backend/doc/openapi.yaml
openapi: 3.0.1
info:
  title: Stupro 2019 Querimonia API
  description: Querimonia ist eine Plattform zum automatischen Beschwerdemanagement. Über die hier beschriebene API kann mit dem Backend kommuniziert werden. Dies beinhaltet Upload von neuen Beschwerden, Abrufen der gespeicherten Beschwerden, Anfragen von Antworttexten, etc.
  version: 0.5.2
servers:
  - url: https://beschwerdemanagement.iao.fraunhofer.de/dev/
  - url: https://beschwerdemanagement.iao.fraunhofer.de/
  - url: https://querimonia.iao.fraunhofer.de/dev
  - url: https://querimonia.iao.fraunhofer.de/
  - url: http://localhost:8080/
...
paths:
...
  /api/responses/{complaintId}:
    parameters:
      - $ref: '#/components/parameters/ComplaintIdParameter'

    get:
      tags:
        - responses
      summary: Gibt die generierte Antwort zu einer Beschwerde zurück.
      description: Zur jeder Beschwerde wird eine Antwort generiert, die über diese Schnittstelle abgerufen werden kann. Eine Antwort enthält eine Liste an Antwortbausteinen sowie eine Liste an Aktionen.
      operationId: getResponse
      responses:
        200:
          description: Die Antwort der Beschwerde wird zurückgegeben.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseSuggestion'
        404:
          description: Es gibt keine Beschwerde mit der gegebenen ID.
...
  /api/responses/plain/{complaintId}:
    parameters:
      - $ref: '#/components/parameters/ComplaintIdParameter'

    get:
      tags:
        - responses
      summary: Gibt die generierte Antwort als String zurück.
      deprecated: true
      description: Gibt den vollen Text einer generierten Antwort als String zurück.  Sollte nur zum Testen verwendet werden.
      operationId: getPlainResponse
      responses:
        200:
          description: Die Antwort der Beschwerde als Gesamttext wird zurückgegeben.
          content:
            text/plain:
              schema:
                type: string
        404:
          description: Es gibt keine Beschwerde mit der gegebenen ID.
  /api/templates:
    get:
      tags:
        - templates
      summary: Gibt die gespeicherten Antwortbausteine zurück.
      description: Hier können alle in der Datenbank gespeicherten Antwortbausteine abgerufen werden. Dabei ist Filtern/Pagination möglich.
      parameters:
        - $ref: '#/components/parameters/CountParameter'
        - $ref: '#/components/parameters/PageParameter'
        - name: sort_by
          in: query
          description: Hier können alle Kategorien angegeben werden, nach denen die Anfrage sortiert werden soll. Die Reihenfolge im Array bestimmt die Priorität beim Sortieren.
          required: false
          schema:
            type: array
            items:
              type: string
              enum: [id_desc, id_asc, priority_desc, priority_asc, name_desc, name_asc]
              default: name_asc
              example: name_asc
        - $ref: '#/components/parameters/KeywordsParameter'

      operationId: getResponseComponents
      responses:
        200:
          description: Die Antwortbausteine werden gefiltert und sortiert zurückgegeben.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ResponseComponent'
        400:
          description: Fehlerhafte Parameter.

    post:
      tags:
        - templates
      summary: Fügt dem System einen neuen Antwortbaustein hinzu.
      description: Fügt einen neuen Antwortbaustein hinzu. Dieser kann Platzhalter enthalten, welche mit Entitäten aus der Beschwerden ersetzt werden.
      requestBody:
        description: Der neue Textbaustein, der dem System hinzufügt werden soll. Im Request-Body müssen ID und requiredEntities nicht angegeben werden, diese werden bei Angabe nicht beachtet.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponseComponent'

      operationId: addResponseComponent
      responses:
        201:
          description: Das Template ist richtig formatiert, die Regelbeschreibung ist gültig und wurde in der Datenbank gespeichert.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseComponent'
        400:
          description: Der Textbaustein enthält ungültige Formatierungen, die Regelbeschreibung ist ungültig oder überschreitet das Zeichenlimit von 5000 Zeichen.
  /api/templates/default:
    post:
      tags:
        - templates
      summary: Fügt dem System eine Menge an Standard-Antwortbausteinen hinzu.
      description: Fügt eine Menge an Standard-Antwortbausteinen hinzu. Diese enthält verschiedene Bausteine für Begrüßungen, Eingangsdaten, Problembeschreibungen, Entschuldigungen, Lösungsvorschläge, Erstattungen und Schlüsse. Die Bausteine gehören zu den Kategorien Fahrer unfreundlich, Fahrt nicht erfolgt, Sonstiges. Jeder Baustein hat drei verschiedene Texte und verschiedene Varianten mit/ohne Platzhaltern.

      operationId: addDefaultComponents
      responses:
        204:
          description: Die Standard Bausteine wurden erfolgreich hinzugefügt.
        500:
          description: Die Datei mit den Standard-Antwortbausteinen ist nicht vorhanden, korrupt oder nicht richtig formatiert.
  /api/templates/{templateId}:
    parameters:
      - name: templateId
        in: path
        description: Die eindeutige ID des Bausteins.
        required: true
        schema:
          type: integer
          format: int32
          minimum: 1
          example: 1

    get:
      tags:
        - templates
      summary: Gibt ein bestimmten Antwortbaustein zurück.
      description: Gibt den Antwortbaustein mit der gegebenen ID zurück.
      operationId: getResponseComponent
      responses:
        200:
          description: Der Antwortbaustein wird zurückgegeben.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseComponent'
        404:
          description: Der Antwortbaustein existiert nicht.

    delete:
      tags:
        - templates
      summary: Löscht einen Antwortbaustein.
      description: Löscht den Antwortbaustein mit der gegebenen ID aus der Datenbank.
      operationId: deleteResponseComponent
      responses:
        204:
          description: Der Antwortbaustein wurde gelöscht.
        404:
          description: Der Antwortbaustein existiert nicht.

    put:
      tags:
        - templates
      summary: Aktualisiert einen bestehenden Antwortbaustein.
      description: Aktualisiert den Antwortbaustein mit der gegebenen ID. Es sollten nur die Properties angegeben werden, die als required markiert sind, "requiredEntities" und ID sollten nicht angegeben werden und werden bei Angabe ignoriert.
      requestBody:
        description: Der aktualisierte Baustein. Im Request-Body müssen ID und requiredEntities nicht angegeben werden, bei Angabe werden diese ignoriert.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResponseComponent'
      operationId: updateResponseComponent
      responses:
        200:
          description: Template erfolgreich aktualisiert.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseComponent'
        404:
          description: Template mit der gegebenen ID ist nicht vorhanden.
        400:
          description: Der Textbaustein enthält ungültige Formatierungen, die Regelbeschreibung ist ungültig oder überschreitet das Zeichenlimit von 5000 Zeichen.
...
components:
  parameters:
    ComplaintIdParameter:
      name: complaintId
      in: path
      description: Die eindeutige ID einer Beschwerde.
      required: true
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 1
...
    CountParameter:
      name: count
      in: query
      description: Anzahl der Elemente, die zurück gegeben werden sollen. Falls nicht angegeben, werden alle Elemente zurückgegeben.
      required: false
      schema:
        type: integer
        format: int32
        minimum: 1
        example: 10
    PageParameter:
      name: page
      in: query
      description: Falls page 0 ist, werden die ersten *count* Elemente zurück gegeben, falls page 1 ist die zweiten *count* Elemente, gibt also die Seitenzahl an. Falls nicht angegeben, wird 0 angenommen. Wird ignoriert, falls *count* nicht gegeben ist.
      required: false
      schema:
        type: integer
        format: int32
        default: 0
        minimum: 0
        example: 0
...
    KeywordsParameter:
      name: keywords
      in: query
      required: false
      description: Bei Angabe werden nur Elemente ausgewertet, die die angebenen Schlüsselwörter enthalten.
      schema:
        type: array
        items:
          type: string
        example:
          - "Busfahrer"
          - "unfreundlich"
...
  schemas:
...
    NamedEntity:
      description: Beschreibt den Anfangs- und Endindex einer *Named-Entity* einer bestimmten Kategorie. Außerdem ist angeben. ob die Entität von einem Nutzer gesetzt wurde und welcher Extraktor zur Findung gefunden wurde. Dieser wird nicht aktualisiert, wenn sich die Extraktoren-Einstellungen ändern! Der gegebene Farbwert, sollte also nur als Fallback verwendet werden.
      type: object
      required:
        - label
        - startIndex
        - endIndex
      properties:
        label:
          type: string
          description: Die Kategorie der Entität.
          example: Datum
        startIndex:
          type: integer
          format: int32
          description: Der Index des Zeichens im Text, mit dem die Named Entity beginnt.
          example: 24
        endIndex:
          type: integer
          format: int32
          description: Der Index des Zeichens im Text, ab dem der Text *nicht* mehr diese Named Entity ist.
          example: 29
        setByUser:
          type: boolean
          description: Gibt an, ob eine Entität vom Nutzer gesetzt wurde (true) oder vom System marktiert wurde (false).
          example: false
        extractor:
          $ref: '#/components/schemas/Extractor'
...
    Extractor:
      description: Ein Extraktor hat einen **eindeutigen** Namen, einen Typ und eine zugeordnete Farbe für jede Entity für die Anzeige im Frontend.
      type: object
      required:
        - name
        - type
        - colors
      properties:
        name:
          type: string
          description: Der Name des Extraktors. Bei KIKuKo Extraktoren muss ein Extraktor mit diesem Namen konfiguriert sein.
          example: GeldbetragExtraktor
        entityLabel:
          type: string
          description: Gibt das Lebel der Entitäten an, die mit diesem Extraktor gefunden werden.
          example: Geldbetrag
        type:
          type: string
          enum: [KIkukoTool]
          description: Eine Bezeichner für den Typ von Entraktor. KIkukoTool wird für Tools von KIkuko verwendet.
          example: KIKuKoTool
        colors:
          type: object
          description: Gibt für beliebige Entity-Labels einen Farbcode an, mit der die Entität dargestellt wird.
          additionalProperties:
            type: string
            format: color
          example:
            "Haltestelle": "#0A2BFF"
            "Name": "#0ABCFF"
    ResponseComponent:
      description: Ein generischer Antwortbaustein. Dieser besteht aus möglicherweise mehereren Alternativtexten, die Platzhalter enthalten können, sowie einer Regelbeschreibung, die angibt, wann der Baustein verwendet werden kann.
      type: object
      required:
        - priority
        - componentName
        - templateTexts
        - rulesXml
      properties:
        componentId:
          type: integer
          format: int32
          description: Die eindeutige ID des Antwortbausteins.
          minimum: 1
          example: 1
        priority:
          type: integer
          format: int32
          example: 100
          description: Eine Zahl, die die Priorität bei der Antwortgenerierung angibt. Bausteine mit hoher Priorität werden zuerst geprüft und ggf. verwendet.
        componentName:
          type: string
          description: Eine **eindeutige** Bezeichung, welchen Teil einer Antwort der Baustein widerspiegelt.
          example: 'Begrüßung'
        templateTexts:
          type: array
          items:
            type: string
            description: Bausteintext mit Platzhaltern.
            example: "Sehr geehrter Herr ${Name},"
        rulesXml:
          type: string
          description: XML-Beschreibung der Regeln, die beschreiben, wann ein Antwortbaustein verwendet werden darf. Muss dem XSD-Schema entsprechen.
        requiredEntities:
          type: array
          items:
            type: string
            example: "Name"
            description: Eine Entity, die für den Antwortbaustein vorhanden sein muss. Muss **nicht** beim Posten eines Bausteins angegeben werden.
    CompletedResponsePart:
      description: Ein Teil der Anwort, der mögliche, asugefüllte Textbausteine enthält.
      type: object
      required:
        - alternatives
        - component
      properties:
        responsePartId:
          type: integer
          format: int32
          minimum: 1
          example: 1
          description: Die eindeutige Id des Antwortteils.
        alternatives:
          type: array
          items:
            $ref: '#/components/schemas/SingleCompletedComponent'
          description: Der Textalternativen des Antwortteils
        component:
          $ref: '#/components/schemas/ResponseComponent'
    SingleCompletedComponent:
      description: Ein einzelner, ausgefüllter Textbaustein mit markierten Entitäten.
      type: object
      required:
        - completedText
        - entities
      properties:
        completedComponentId:
          type: integer
          format: int32
          minimum: 1
          example: 1
          description: Eine eindeutige Id des ausgefüllten Bausteins.
        completedText:
          type: string
          description: Der ausgefüllte Text des Bausteins.
          example: "Guten Tag, Max Mustermann,\nVielen Dank für Ihre Nachricht am 23.01.1999."
        entities:
          type: array
          items:
            $ref: '#/components/schemas/NamedEntity'
          description: Enthält alle Entitäten, die in der Antwort vorhanden sind.
    ResponseSuggestion:
      description: Vollständiger Antwortvorschlag, bestehend aus ausgefüllten Text-Komponenten und Aktionen.
      type: object
      required:
        - components
        - actions
      properties:
        components:
          type: array
          items:
            $ref: '#/components/schemas/CompletedResponsePart'
        actions:
          type: array
          items:
            $ref: '#/components/schemas/Action'
    Action:
      description: Eine Aktion kann automatisch beim Abschluss einer Beschwerde ausgelöst werden, wenn bestimmt Regeln dafür erfüllt sind. Zum Beispiel können E-Mails versendet werden, Gutscheine als Attachment angehängt werden, Beschwerden weitergeleitet werden etc.
      type: object
      required:
        - name
        - actionCode
        - rulesXml
        - parameters
      properties:
        actionId:
          description: Die eindeutige ID der Aktion.
          type: integer
          format: int32
          minimum: 1
          example: 1
        name:
          description: Ein eindeutiger Bezeichnert für die Aktion.
          type: string
          example: "GutscheinFreifahrt"
        actionCode:
          $ref: '#/components/schemas/ActionCode'
        rulesXml:
          description: Die Regeln, wann eine Aktion ausgelöst werden kann im XML-Format.
          type: string
          format: xml
          example: '<Rules><Subject value="Fahrt nicht erfolgt"/></Rules>'
        parameters:
          description: Zusätzliche Parameter für die Aktion, zum Beispiel an wen eine E-Mail gesendet wird oder den Wert des Gutscheins.
          type: object
          additionalProperties:
            type: string
          example:
            "E-Mail": "beispiel@gmail.com"
            "Wert": "5"
    ActionCode:
      description: Eine Bezeichung, die den Typ der Aktion angibt.
      type: string
      enum:
        - SEND_MAIL
        - ATTACH_VOUCHER
      example: "ATTACH_VOUCHER"
...
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic